Refactor duplicate code in ManagedResoures
