+++ src/QueryType/ManagedResources/Query/Command/AbstractCreate.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Solarium\QueryType\ManagedResources\Query\Command;
+
+use Solarium\Core\Client\Request;
+use Solarium\QueryType\ManagedResources\Query\AbstractCommand;
+
+abstract class AbstractCreate extends AbstractCommand
+{
+    /**
+     * Returns request method.
+     *
+     * @return string
+     */
+    public function getRequestMethod(): string
+    {
+        return Request::METHOD_PUT;
+    }
+
+    /**
+     * Empty.
+     *
+     * @return string
+     */
+    public function getTerm(): string
+    {
+        return '';
+    }
+}
