+++ src/QueryType/ManagedResources/Query/Stopwords/Command/Remove.php
@@ -2,11 +2,10 @@
 
 namespace Solarium\QueryType\ManagedResources\Query\Stopwords\Command;
 
-use Solarium\Core\Client\Request;
-use Solarium\QueryType\ManagedResources\Query\AbstractCommand;
+use Solarium\QueryType\ManagedResources\Query\Command\AbstractRemove;
 use Solarium\QueryType\ManagedResources\Query\Stopwords;
 
-class Remove extends AbstractCommand
+class Remove extends AbstractRemove
 {
     /**
      * Returns command type, for use in adapters.
@@ -16,33 +15,5 @@ class Remove extends AbstractCommand
     public function getType(): string
     {
         return Stopwords::COMMAND_REMOVE;
-    }
-
-    /**
-     * Returns request method.
-     *
-     * @return string
-     */
-    public function getRequestMethod(): string
-    {
-        return Request::METHOD_DELETE;
-    }
-
-    /**
-     * Returns the raw data to be sent to Solr.
-     */
-    public function getRawData(): string
-    {
-        return '';
-    }
-
-    /**
-     * Empty.
-     *
-     * @return string
-     */
-    public function getTerm(): string
-    {
-        return '';
     }
 }
