+++ src/QueryType/ManagedResources/Query/Synonyms/Command/Delete.php
@@ -2,19 +2,11 @@
 
 namespace Solarium\QueryType\ManagedResources\Query\Synonyms\Command;
 
-use Solarium\Core\Client\Request;
-use Solarium\QueryType\ManagedResources\Query\AbstractCommand;
+use Solarium\QueryType\ManagedResources\Query\Command\AbstractDelete;
 use Solarium\QueryType\ManagedResources\Query\Synonyms;
 
-class Delete extends AbstractCommand
+class Delete extends AbstractDelete
 {
-    /**
-     * Term to be deleted.
-     *
-     * @var string
-     */
-    protected $term = '';
-
     /**
      * Returns command type, for use in adapters.
      *
@@ -23,48 +15,5 @@ class Delete extends AbstractCommand
     public function getType(): string
     {
         return Synonyms::COMMAND_DELETE;
-    }
-
-    /**
-     * Returns request method.
-     *
-     * @return string
-     */
-    public function getRequestMethod(): string
-    {
-        return Request::METHOD_DELETE;
-    }
-
-    /**
-     * Empty.
-     *
-     * @return string
-     */
-    public function getRawData(): string
-    {
-        return '';
-    }
-
-    /**
-     * Returns the term to be deleted.
-     *
-     * @return string
-     */
-    public function getTerm(): string
-    {
-        return $this->term;
-    }
-
-    /**
-     * Set the term to be deleted.
-     *
-     * @param string $term
-     *
-     * @return self
-     */
-    public function setTerm(string $term): self
-    {
-        $this->term = $term;
-        return $this;
     }
 }
