+++ src/QueryType/ManagedResources/Query/Synonyms/Command/Remove.php
@@ -2,11 +2,10 @@
 
 namespace Solarium\QueryType\ManagedResources\Query\Synonyms\Command;
 
-use Solarium\Core\Client\Request;
-use Solarium\QueryType\ManagedResources\Query\AbstractCommand;
+use Solarium\QueryType\ManagedResources\Query\Command\AbstractRemove;
 use Solarium\QueryType\ManagedResources\Query\Synonyms;
 
-class Remove extends AbstractCommand
+class Remove extends AbstractRemove
 {
     /**
      * Returns command type, for use in adapters.
@@ -16,33 +15,5 @@ class Remove extends AbstractCommand
     public function getType(): string
     {
         return Synonyms::COMMAND_REMOVE;
-    }
-
-    /**
-     * Returns request method.
-     *
-     * @return string
-     */
-    public function getRequestMethod(): string
-    {
-        return Request::METHOD_DELETE;
-    }
-
-    /**
-     * Returns the raw data to be sent to Solr.
-     */
-    public function getRawData(): string
-    {
-        return '';
-    }
-
-    /**
-     * Empty.
-     *
-     * @return string
-     */
-    public function getTerm(): string
-    {
-        return '';
     }
 }
